requirements.txt
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.0
pillow==10.4.0
pyyaml==6.0.2
tqdm==4.66.4
